+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Compatibility:
 * Fix issue with `Float#round` when `self` is `-0.0`.
 * Fix `String#unpack` issue with `m0` format (#2065).
 * Fix issue with `File.absolute_path` returning a path to current directory (#2062).
+* Update `Range#cover?` to handle `Range` parameter.
 
 Performance: